Change the return values
SBI needs to distinguish return values for the SBI calling sequence itself, and the values from the functions called. I think that we should reserve a0 for SBI errors, and a1 for the return value of functions themselves. Otherwise, as we've found in coreboot, there's not a good clean way to make console_putchar optionally compiled in; further, in using -38 for the error return from SBI, you will eternally mix up that one bit pattern with the bit pattern from all possible future SBI functions. Yuck.

This is how Unix originally worked: there were actually multiple return values from a system call (one of them being the PDP-11 carry bit, but still ...). It was only when we had to make the move to C, a single-return-value language, that we had to go with a single return value from system calls! And even in that case, there really are more than one: that's what errno is for: it's the "second return value."

Save yourself future pain and get this right now. Think of the call to SBI as you would an RPC, and allow that call to return errors; then allow the functions themselves to return an error distinct from possible SBI errors. You won't regret it.