Define some terminology
Terms like should et. al. are commonly understood to introduce a
requirement, but it is sometimes unclear exactly what they mean. Link to
the RFC specifying particular meanings for those terms. This may cause
the meaning of some parts of the spec to change if these terms were
previously used with different intended or implied meanings.

The term "hart" is used extensively without definition. The idea of an
EEI is used in the following commit "sbi: Specify the instruction set".

What exactly constitutes a valid **hartid** was not previously
explicitly defined. I have taken a broad view here, and tried not to
constrain valid hartids past what is reasonable. A more narrow
interpretation could allow more efficient implementations in some
instances.

Signed-off-by: Sean Anderson <seanga2@gmail.com>