Specify the initial state
The initial state of software running in the SBI was not specified.
Particularly egregious was the omission of the requirement that a0
contain the hartid of the hart. That behavior was used to justify not
including a function to get the current hartid (e.g. in #25), so it
should be specified.

SBI is intended as an interface for supervisor mode software, however,
there is no technical reason why it cannot also be a "UBI." This mode of
operation is currently supported by OpenSBI. I think it should be
captured by this specification. There is a slight conflict with the HSM
extension, in that harts are requires to start in S-mode. If a system
only supports U-mode, then it would be impossible to start harts with
the HSM extension, as it is currently written. If it is undesired to
reference U-mode, then those references can be removed (or split off
into a separate series).

This also specifies the return value of sbi_hart_status. This is to
avoid a "slow start" situation where a hart is part of an execution
environment, but starts much later than another hart. This could cause
confusing behavior where it appeared to be stopped, and then start
without a call to sbi_start_hart. Perhaps this sort of behavior should
be explicitly disallowed by the HSM extension?

Signed-off-by: Sean Anderson <seanga2@gmail.com>