intro: Re-word the "you can't implement part of an extension" bit
This one was just a bit awkward: trying to say "implemented" and "in
total" sort of paints us into a corner with some of the optional
functionality in the spec -- for example, some SBI implementations may
just ignore the address fields in the fences and always do global
fences.

The new langeage is just a bit more precise, but IMO it's what the old
language was trying to say.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>