HSM: Clarify that firmware must set M-mode state before transitioning
We don't really care if firmware executes S-mode instructions, which may
be a useful thing to do (maybe it wants to leave a stub to avoid
transitions).  We just care that the relevant state is set up before we
end up back in the supervisor software, this clears up the wording.

Signed-off-by: Palmer Dabbelt <palmerdabbelt@google.com>