Clarify invalid-param and not-supported errors
SBI_ERR_INVALID_PARAM is returned for type parameters when the type
value is unknown to the SBI implementation. For standard types,
unknown values are considered reserved. For platform-specific types,
unknown values imply the platform has not implemented it, but the
platform reserves the right to do so. In both cases the values are
"reserved", hence the same error for both. In contrast,
SBI_ERR_NOT_SUPPORTED is returned when the type value is known, but
at least one dependency is missing which prohibits the platform from
supporting it. Clarify these two errors by including text about
platform-specific types.

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>
Reviewed-by: Conor Dooley <conor.dooley@microchip.com>