Introduce CPPC extension
ACPI defines the Collaborative Processor Performance Control (CPPC),
a mechanism which is an abstract and flexible mechanism for the
Operating System to collaborate with an entity in the platform to
manage the performance of the processors.

The SBI CPPC extension provides an abstraction to access CPPC
registers through SBI calls.

Reviewed-by: Andrew Jones <ajones@ventanamicro.com>
Reviewed-by: Atish Patra <atishp@rivosinc.com>
Signed-off-by: Sunil V L <sunilvl@ventanamicro.com>