SUSP: Improve suspend scope text
The intention of the changed sections is to state that system suspend
might not actually suspend the entire system, but rather the view of
the system available to the S-mode software domain executing the call.
Address the intention more directly. The alternate language also avoids
inadvertently constraining implementations that might implement SBI
calls in hardware or microcode. Finally, there is no guarantee even in
the native case that any power transition may take place.

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>