Spell hart consistently
There are instances of HART(s), hart(s), and Hart(s) throughout the
specification. Convert all HART(s) instances to hart(s). After this
patch there are no more instances of HART(s) and Hart(s) only occurs
when used as the first word of a sentence of within a title.

Signed-off-by: Andrew Jones <ajones@ventanamicro.com>