Remove non-member contributions
As per the RVI policy, remove the non-member entries from the contributor
list as determined by the RVI tool. The information may be stale and
If the authors are a member now or wish to be a RVI member, please
reach out to help@riscv.org. We would be happy to add them to the contributor
list.

Signed-off-by: Atish Patra <atishp@rivosinc.com>