Add Firmware Features extension
This extension is meant to provide a way for supervisor-mode software to
request changes of some feature handling provided by the M-mode
firmware. Two new functions are added to get and set theses feature with
specific values. This first FWFT extension version documents the
following features:
- MISALIGNED_EXC_DELEG: Delegate misaligned access exception
- SOFTWARE_CHECK_EXC_DELEG: Delegate software check exception for
  supervisor-mode[1]
- LANDING_PAD: Control Landing Pad for supervisor-mode[1]
- SHADOW_STACK: Control Shadow Stack for supervisor-mode [1]
- DOUBLE_TRAP: Control Double Trap support for supervisor-mode [2]
- PTE_AD_HW_UPDATING: Control hardware updating of PTE A/D bits

An openSBI implementation is visible at [3] as well a kvm-unit-tests
with MISALIGNED_EXC_DELEG testing at [4].

Link: https://github.com/riscv/riscv-cfi [1]
Link: https://github.com/ved-rivos/riscv-double-trap [2]
Link: https://github.com/rivosinc/opensbi/tree/dev/cleger/fwft [3]
Link: https://github.com/clementleger/kvm-unit-tests/tree/dev/cleger/fwft [4]
Signed-off-by: Clément Léger <cleger@rivosinc.com>
Reviewed-by: Samuel Holland <samuel.holland@sifive.com>