Add a new function for event discovery
The supervisor software may need to be aware of which standard or
raw events are allowed by the platform. This will help the supervisor
software to advertise only those events instead of all standard events.
In addition to that, it may just waste time trying to match a counter
for each invalid event at runtime which may lead to undesirable results.
Currently, this problem is solved by invoking CFG_MATCH call for each
event which is not ideal as it may increase the bootime with unnecessary
SBI calls at boot time.

Introduce a new function that allows to discover support events with a
single SBI call which is very efficient.

Signed-off-by: Atish Patra <atishp@rivosinc.com>