Clarify sbi_pmu_event_get_info() error conditions
For compatibility with event encodings defined by a future version of
this standard, a call to sbi_pmu_event_get_info() should not fail when
an unknown or reserved `event_idx` value is provided in the shared
memory. The SBI implementation should simply report that the event is
unsupported in the corresponding output word.

Clarify that the condition for returning SBI_ERR_INVALID_PARAM only
applies to the reserved bits in the `event_idx` word, and not the value
of the 20-bit `event_idx` field.

Reviewed-by: Atish Patra <atishp@rivosinc.com>
Signed-off-by: Samuel Holland <samuel.holland@sifive.com>