ext-sse: add SBI_ERR_NOT_SUPPORTED return value
Some events might only be supported if there is hardware support for
them (such as the PMU overflow one for instance). In that case,
returning SBI_ERR_NOT_SUPPORTED would allow to differentiate between
valid events but not supported and non valid events in which case
SBI_ERR_INVALID_PARAM is returned.

Reviewed-by: Atish Patra <atishp@rivosinc.com>
Signed-off-by: Clément Léger <cleger@rivosinc.com>