ext-fwft: use the provided value in fwft_set(POINTER_MASKING_PMLEN)
Using a greater value without supervisor approval is problematic,
because it could result in behavior that the supervisor cannot handle.

We've discussed an extra function with detection bitmask for this, but
decided to have the supervisor software loop over all possible values
and find one that works.  Luckily there are only two options right now.

There was also an attempt to add a note with a recommendation to
supervisor software writers, but the resulting discussion failed to
produce a satisfactory formulation and the fix is needed ASAP.

Signed-off-by: Radim Krčmář <rkrcmar@ventanamicro.com>